+++ packages/core/src/types/config.ts
@@ -1,7 +1,7 @@
-import type { ParseMitosisOptions } from '../parsers/jsx/types';
-import { targets } from '../targets';
+import type { ParseMitosisOptions } from '@/parsers/jsx';
+import { targets } from '@/targets';
 import type { MitosisComponent } from './mitosis-component';
-import { BaseTranspilerOptions } from './transpiler';
+import { BaseTranspilerOptions, TranspilerGenerator } from './transpiler';
 export type Format = 'esm' | 'cjs';
 export type Language = 'js' | 'ts';
 interface TranspilerOptions {
@@ -19,8 +19,22 @@ export type generatorsOption = {
   [K in Target]: NonNullable<Targets[K]>;
 };
 
+export interface TargetContext {
+  target: Target;
+  generator: TranspilerGenerator<NonNullable<MitosisConfig['options'][Target]>>;
+  outputPath: string;
+}
+
+export interface OutputFiles {
+  outputDir: string;
+  outputFilePath: string;
+}
+
 export type MitosisConfig = {
   generators?: generatorsOption;
+  /**
+   * Apply common options to all targets
+   */
   commonOptions?: Omit<BaseTranspilerOptions, 'experimental'>;
   /**
    * List of targets to compile to.
@@ -60,6 +74,7 @@ export type MitosisConfig = {
    *   react: {
    *     stateType: 'builder';
    *     stylesType: 'styled-jsx'
+   *     plugins: [myPlugin]
    *   }
    * }
    * ```