+++ packages/core/src/types/plugins.ts
@@ -1,16 +1,30 @@
+import { OutputFiles, TargetContext } from '@/types/config';
 import { MitosisComponent } from './mitosis-component';
 
-export type Plugin = (options?: any) => {
-  json?: {
-    // Happens before any modifiers
-    pre?: (json: MitosisComponent) => MitosisComponent | void;
-    // Happens after built in modifiers
-    post?: (json: MitosisComponent) => MitosisComponent | void;
-  };
-  code?: {
-    // Happens before formatting
-    pre?: (code: string, json: MitosisComponent) => string;
-    // Happens after formatting
-    post?: (code: string, json: MitosisComponent) => string;
-  };
+export type MitosisBuildPlugin = (
+  targetContext: TargetContext,
+  files?: {
+    componentFiles: OutputFiles[];
+    nonComponentFiles: OutputFiles[];
+  },
+) => void | Promise<void>;
+
+export type MitosisJsonPlugin = (json: MitosisComponent) => MitosisComponent | void;
+
+export type MitosisCodePlugin = (code: string, json: MitosisComponent) => string;
+
+export type MitosisHook<T> = {
+  pre?: T;
+  post?: T;
+};
+
+export type MitosisPlugin = (options?: any) => {
+  name?: string;
+  order?: number;
+  // Happens before/after build
+  build?: MitosisHook<MitosisBuildPlugin>;
+  // Happens before/after any modifiers
+  json?: MitosisHook<MitosisJsonPlugin>;
+  // Happens before/after formatting
+  code?: MitosisHook<MitosisCodePlugin>;
 };